# backend/equis/src/main/java/es/urjc/daw/equis/controller/UserController.java
package es.urjc.daw.equis.controller;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.ResponseBody;

import es.urjc.daw.equis.service.UserService;

@Controller
public class UserController {

    @Autowired
    private UserService userService;

    public UserController(UserService userService) {
        this.userService = userService;
    }

    @GetMapping("/users/{id}/profile-picture")
    @ResponseBody
    public ResponseEntity<byte[]> getProfilePicture(@PathVariable Long id) throws Exception {

        byte[] image = userService.getProfilePictureBytes(id);

        if (image == null) {
            return ResponseEntity.notFound().build();
        }

        return ResponseEntity.ok()
                .contentType(MediaType.IMAGE_JPEG)
                .body(image);
    }
}